+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 - `limestone/Image` props `backgroundColor` and `backgroundSrc` to be able to change the background of a transparent image
 - `limestone/ImageItem` props `backgroundColor` and `backgroundSrc` to be able to change the background of a transparent image
+- `limestone/Slider` props `ticks` and `labels` to display equally spaced tick marks and labels
+- `limestone/Slider` prop `alignStepsWithTicks` to snap the knob to tick marks
 
 ### Changed